third_party/WebKit/LayoutTests/editing/deleting/merge-paragraph-containing-noneditable.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => assert_selection(
  [
    '<div contenteditable>',
      '<p>First Paragraph.</p>',
      '<p>',
        '|First',
        '<span class="noneditable" contenteditable="false">Second</span>',
        'Third',
      '</p>',
    '</div>'
  ].join(''),
  'delete',
  [
    '<div contenteditable>',
      '<p>',
        'First Paragraph.|First',
        '<span class="noneditable" contenteditable="false">Second</span>',
        'Third',
      '</p>',
    '</div>'
  ].join('')),
  'Merging paragraphs with mixed editability does not cause loss of elements.');
</script>
